Change sans-serif to Lato, add Georgia to serif
Open Sans has a sort of overbearing 'safeness' about it. Lato is more
modern whilst still being safe and supporting the same weights. Kept
Open Sans as a fallback, but without loading as an external font. Also
added Georgia as the fallback for serif, and also added Baskerville as
that seems the closest Mac default.