Add custom styles for Gutenberg galleries
Custom styles and scripts for displaying Gutenberg-generated galleries.
Items are displayed in a flexbox grid, and if JS is available the first
gallery on the screen has its display controlled by Masonry. This
controls reflowing when enlarging any particular gallery item.

The script will need refactoring to allow use on multiple galleries per
page. As Masonry is loaded as a library, I will probably also use it for
grids of images on archive pages also.